# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Event Management Platform API
  version: 0.1.0
  description: |
    Domain API only. Sign-up and sign-in are owned by Better Auth and are not
    in this spec.

    Partial contract: only the routes that exist today. GET /events/{id} is
    role-dependent. The same Grant.Can check that allows the request also
    decides which fields are present.

    PATCH /sessions/{id} is the first write. ROOM_CONFLICT and STALE_VERSION
    are both HTTP 409; the client branches on code, not status.

servers:
  - url: http://localhost:8080

security: []

paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Liveness check
      description: Open. No session cookie required.
      responses:
        "200":
          description: Process is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  /me:
    get:
      operationId: getMe
      summary: Current user
      description: |
        The caller identified by the Better Auth session cookie.
        Returns that person's own id, name, and email.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/Internal"

  /events:
    get:
      operationId: listEvents
      summary: List events the caller can see
      description: |
        Keyset on `id` (uuidv7, time-ordered). No offset. Only events the
        caller is a member of with event.read.

        List items use PresentEvent with no roster loaded, so members and
        emails are never in this response.
      security:
        - sessionCookie: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque. From nextCursor of the previous page.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: One page of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/Internal"
    post:
      operationId: createEvent
      summary: Create an event
      description: |
        Any signed-in user may create an event. Creation is gated on
        authentication only: there is no event.create permission, because
        per-event roles cannot predate the event. The creator is inserted as
        admin on the new event.

        timeZone must be a name in time_zones. startsAt and endsAt are local
        wall-clock in that zone, parsed by tz.Instant.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreate"
      responses:
        "201":
          description: Created event. Roster omitted on this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/Internal"

  /events/{id}:
    get:
      operationId: getEvent
      summary: Get one event
      description: |
        Returns a single event the caller belongs to.

        The grant is checked before the event row. A caller who is not a
        member gets FORBIDDEN whether the event exists or not. NOT_FOUND is
        only returned when the caller has a grant and the row is gone.

        The response body is role-dependent. PresentEvent uses the same
        Grant.Can check as the URL:

        - member.read — include members. Without it the key is omitted
          (attendee). Go omitempty: the field is absent, not null.
        - user.email.read — include email on each member. Without it each
          email field is omitted, not null.

        Admin and contributor currently hold both permissions. Attendee
        holds neither, so the attendee body is the event fields only.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Event. Roster and emails only when the grant allows them.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              examples:
                admin:
                  summary: Admin or contributor — roster and emails present
                  value:
                    id: "01a00fb3-c792-76a7-9336-6d5ef6a9577a"
                    name: Test Conf
                    description: Just a conference
                    timeZone: Europe/Berlin
                    startsAt: "2026-03-28T19:00:00Z"
                    endsAt: "2026-03-30T16:00:00Z"
                    members:
                      - userId: 0fXEL014POMXqTQolyl1IvbdA4uZ0PD0
                        name: Admin Ada
                        email: admin@example.com
                        role: admin
                attendee:
                  summary: Attendee — no members key, no emails
                  value:
                    id: "01a00fb3-c792-76a7-9336-6d5ef6a9577a"
                    name: Test Conf
                    description: Just a conference
                    timeZone: Europe/Berlin
                    startsAt: "2026-03-28T19:00:00Z"
                    endsAt: "2026-03-30T16:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"

  /events/{eventId}/sessions:
    get:
      operationId: listEventSessions
      summary: List sessions of one event
      description: |
        Ordered by startsAt, then id. Includes room, times, and version for
        the schedule view. Needs session.read.

        Grant first, same as GET /events/{id}: not a member is FORBIDDEN
        whether the event exists or not. NOT_FOUND only if the caller has a
        grant and the row is gone.
      security:
        - sessionCookie: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sessions in start order
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
    post:
      operationId: createEventSession
      summary: Create a session
      description: |
        startsAt and endsAt are local wall-clock in the event's zone, parsed
        by tz.Instant. Room overlap is the exclusion constraint (23P01 →
        ROOM_CONFLICT), not a check-then-insert. Needs session.create.

        Grant first, same as GET /events/{id}: not a member is FORBIDDEN
        whether the event exists or not.
      security:
        - sessionCookie: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
      responses:
        "201":
          description: Created session, including version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: ROOM_CONFLICT — exclusion constraint rejected the write
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: ROOM_CONFLICT
                reason: room is taken
                conflict:
                  conflictingSessionId: "019b76da-d46d-7000-8000-00002c6d0000"
                  conflictingTitle: Talk A2
        "500":
          $ref: "#/components/responses/Internal"

  /events/{eventId}/rooms:
    get:
      operationId: listEventRooms
      summary: List rooms of one event
      description: |
        Needs room.read. Grant first, same as GET /events/{id}: not a member
        is FORBIDDEN whether the event exists or not.
      security:
        - sessionCookie: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rooms
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
    post:
      operationId: createEventRoom
      summary: Create a room
      description: |
        Needs room.manage. Name is unique per event. Grant first, same as
        GET /events/{id}: not a member is FORBIDDEN whether the event exists
        or not.
      security:
        - sessionCookie: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomCreate"
      responses:
        "201":
          description: Created room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"

  /events/{eventId}/invitations:
    get:
      operationId: listEventInvitations
      summary: List invitations of one event
      description: |
        Keyset on `id` (uuidv7, time-ordered). No offset. Index is
        invitations_event_id_id_idx on (event_id, id). ADR 0004.

        Needs invitation.read. Grant first, same as GET /events/{id}: not a
        member / no invitation.read is FORBIDDEN whether the event exists or
        not. Attendees cannot page invitations.

        Invitee email uses the same Grant.Can as PresentEvent: omitted when
        the grant lacks user.email.read. Not sent as null — Go omitempty.
      security:
        - sessionCookie: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque. From nextCursor of the previous page. Same encoding as GET /events.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: One page of invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"

  /sessions/{id}:
    patch:
      operationId: patchSession
      summary: Update a session's time and/or room
      description: |
        Partial update. A missing field is left unchanged. `roomId: null`
        unschedules the session (no room). `title`, `description`, `startsAt`,
        and `endsAt` are not nullable — omit them to leave them.

        `version` is required. The UPDATE matches that version and then
        increments. A mismatch is STALE_VERSION, not a silent overwrite.

        `startsAt` / `endsAt` are local wall-clock times in the event's IANA
        zone, with no offset. The server parses them with tz.Instant.
        A spring-forward gap or a fall-back fold is VALIDATION_ERROR.

        Room double-booking is the database exclusion constraint, not a
        SELECT-then-UPDATE. The constraint error is mapped to ROOM_CONFLICT.

        Needs session.update on the event. Grant first, same as GET
        /events/{id}: a caller without session.update gets FORBIDDEN
        whether the session exists or not. NOT_FOUND is only returned
        when the caller has a grant and the row is gone.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionPatch"
            examples:
              reschedule:
                summary: Move time, keep the room
                value:
                  version: 1
                  startsAt: "2026-03-08T12:00:00"
                  endsAt: "2026-03-08T13:00:00"
              unschedule:
                summary: Clear the room
                value:
                  version: 1
                  roomId: null
              titleOnly:
                summary: Missing fields are left unchanged
                value:
                  version: 1
                  title: Keynote
      responses:
        "200":
          description: Updated session, including the new version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Either the room is taken or the version is stale. Same HTTP
            status. Distinct `code`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                roomConflict:
                  summary: ROOM_CONFLICT — exclusion constraint rejected the write
                  value:
                    code: ROOM_CONFLICT
                    reason: room is taken
                    conflict:
                      conflictingSessionId: "019b76da-d46d-7000-8000-00002c6d0000"
                      conflictingTitle: Talk A2
                staleVersion:
                  summary: STALE_VERSION — includes current state, no second GET
                  value:
                    code: STALE_VERSION
                    reason: version has changed
                    conflict:
                      currentVersion: 2
                      currentState:
                        id: "019b76da-d46d-7000-8000-00002c6d0001"
                        eventId: "019b76da-d44e-7000-8000-00002c4e0000"
                        roomId: "019b76da-d46d-7000-8000-00002c6d0002"
                        title: Talk A1
                        description: ""
                        startsAt: "2026-03-08T13:00:00Z"
                        endsAt: "2026-03-08T14:00:00Z"
                        version: 2
        "500":
          $ref: "#/components/responses/Internal"

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Signed Better Auth session cookie. Go looks up auth.sessions.

  responses:
    Unauthenticated:
      description: No cookie, bad signature, or expired session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: UNAUTHENTICATED
            reason: not signed in
    Forbidden:
      description: Signed in, but not a member, or lacking the permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: FORBIDDEN
            reason: not allowed
    NotFound:
      description: Event or session id does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: NOT_FOUND
            reason: event not found
    ValidationError:
      description: |
        Bad input. Includes a nonexistent or ambiguous local time from
        tz.Instant. Offending fields are in fieldErrors.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: VALIDATION_ERROR
            reason: nonexistent local time
            fieldErrors:
              - field: startsAt
                reason: nonexistent local time
    Internal:
      description: Unexpected failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INTERNAL
            reason: grant lookup failed

  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, reason]
      properties:
        code:
          type: string
          description: Machine-readable. The client branches on this, never on reason.
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL
            - VALIDATION_ERROR
            - ROOM_CONFLICT
            - STALE_VERSION
        reason:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        conflict:
          $ref: "#/components/schemas/Conflict"

    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string

    Conflict:
      type: object
      additionalProperties: false
      properties:
        conflictingSessionId:
          type: string
          format: uuid
        conflictingTitle:
          type: string
        currentVersion:
          type: integer
        currentState:
          $ref: "#/components/schemas/Session"

    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          const: ok

    Me:
      type: object
      additionalProperties: false
      required: [id, name, email]
      properties:
        id:
          type: string
          description: Better Auth user id. Text, not uuid.
        name:
          type: string
        email:
          type: string
          format: email

    Member:
      type: object
      additionalProperties: false
      required: [userId, name, role]
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
          description: |
            Omitted when the grant lacks user.email.read. Not sent as null —
            Go omitempty drops an empty string.
        role:
          type: string
          enum: [admin, contributor, attendee]

    Event:
      type: object
      additionalProperties: false
      required: [id, name, description, timeZone, startsAt, endsAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        timeZone:
          type: string
          description: IANA zone name stored on the event
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
          description: |
            Omitted when the grant lacks member.read (attendee). Not sent as
            null — Go omitempty drops a nil slice.

    Session:
      type: object
      additionalProperties: false
      required: [id, eventId, title, description, startsAt, endsAt, version]
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        roomId:
          type: ["string", "null"]
          format: uuid
          description: Null when the session has no room.
        title:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
          description: Instant, RFC 3339. Local wall clock is this plus the event zone.
        endsAt:
          type: string
          format: date-time
        version:
          type: integer

    SessionPatch:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: integer
          description: Version the client last read. Must match or the write is STALE_VERSION.
        title:
          type: string
          description: Omit to leave unchanged. Not nullable.
        description:
          type: string
        roomId:
          type: ["string", "null"]
          format: uuid
          description: |
            Omit to leave unchanged. JSON null clears the room. A uuid assigns
            a room of this event.
        startsAt:
          type: string
          description: |
            Local wall-clock in the event's IANA zone, no offset
            (`2026-03-08T09:00:00`). Parsed by tz.Instant.
        endsAt:
          type: string
          description: Same shape as startsAt.

    EventPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        nextCursor:
          type: string
          description: Opaque. Omit when this is the last page.

    EventCreate:
      type: object
      additionalProperties: false
      required: [name, timeZone, startsAt, endsAt]
      properties:
        name:
          type: string
        description:
          type: string
        timeZone:
          type: string
          description: IANA name. Must exist in time_zones.
        startsAt:
          type: string
          description: Local wall-clock in timeZone, no offset. tz.Instant.
        endsAt:
          type: string

    SessionCreate:
      type: object
      additionalProperties: false
      required: [title, startsAt, endsAt]
      properties:
        title:
          type: string
        description:
          type: string
        roomId:
          type: ["string", "null"]
          format: uuid
        startsAt:
          type: string
          description: Local wall-clock in the event zone, no offset. tz.Instant.
        endsAt:
          type: string

    Room:
      type: object
      additionalProperties: false
      required: [id, eventId, name, capacity]
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        name:
          type: string
        capacity:
          type: integer

    RoomCreate:
      type: object
      additionalProperties: false
      required: [name, capacity]
      properties:
        name:
          type: string
        capacity:
          type: integer
          minimum: 1

    Invitation:
      type: object
      additionalProperties: false
      required: [id, eventId, role, status]
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: |
            Invitee email. Omitted when the grant lacks user.email.read.
            Not sent as null — Go omitempty drops an empty string.
        role:
          type: string
          enum: [admin, contributor, attendee]
        status:
          type: string
          enum: [pending, accepted, declined, revoked]
        invitedBy:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
          description: Set when the invitee has an account (accepted).

    InvitationPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Invitation"
        nextCursor:
          type: string
          description: Opaque. Omit when this is the last page. Same encoding as GET /events.